feat: refactor the MobilePortStream class
